Add ability to limit minimization
This commit introduces a new commandline option, --minimizationLimit=N, which makes it possible to
specify the minimium number of instructions for programs that are stored into the corpus. This can
help avoid "over-minimization" of corpus samples which can negatively impact the fuzzer's
performance if features of a program are removed during minimization which could later be mutated to
cause interesting behaviour or crashes.