A new type system for Fuzzilli
This commit consists of three different parts:

1. A new type system (see FuzzIL/TypeSystem.swift). The goal of this type system
is to be as simple as possible while still being able to express all (relevant)
operations that can be performed on a JS value. As such, the type system e.g.
has to be able to express that a value can be called as a function or
constructed with the 'new' keyword, or that it is an object with a set of
properties and methods. It also has to express that something is e.g. both an
object with methods and properties and a function that can be called. Finally,
it has to be able to express that a variable can be one of several types (e.g.
because the variable is written to again, or because different properties are
assigned in conditionally executing branches).
The type system is based on union types, with a notion of "merged" types to
express that a value is two or more types at the same time.

2. A static model of the JS environment (see Core/JavaScriptEnvironment.swift).
This model stores the types of all (relevant) entities that are available in a
default JS environment: builtin objects, constructors, functions etc. as well as
property types and method signatures.

3. An abstract intepreter that can perform lightweight interpretation of FuzzIL
code to compute static type information. For that purpose it interacts with the
environment to obtain type information for various builtins.

This change should allow code generators to generate better code as they now
have more information about e.g. available methods or properties at hand. On the
other hand, the new features can be disabled by setting
Configuration.useAbstractInterpretation to false and thus turning off most of
the new logic of the abstract interpreter. This will leave most variable types
at .unknown, thus "emulating" the old behaviour.