Now exposing only a single function, fuzzilli(), from the JS engines
Previously, we exposed a crash and a __fuzzout__ function in every engine. The
problem with that is that fuzzilli quite frequently discovered the crash
function (e.g. by enumerating the global object) and thus produced various
invalid crashes. With this patch, only a single function, fuzzilli(), is exposed
in every engine. This function expects a string as first argument that is used
to select the operation to perform (crash or print to the fuzzilli output
stream). Fuzzilli will then (hopefully) not be able to "guess" these magic
strings, in which case the fuzzilli function is harmless to execute.