Add diagnostics mode
Add a diagnostics mode to save executions that failed or timed-out.
This allows analysis of the output of the fuzzer.
Diagnostics mode also saves all executions made on a REPRL instance
to find scripts that may corrupt the state of the REPRL and saves all
executed scripts if it detects such a REPRL failure.