Programs are now immutable (#138)
As programs are reference counted, it becomes much easier to reason about
them if they are immutable. Moreover, it is unclear how to deal with runtime
type information if a program is modified in place.

This commit introduces a new struct, Code, which is simply a list of
instructions that can be modified (if the struct itself is mutable). A
Program now contains an immutable code instance to hold its instructions.
The separation simplifies a few things as code that previously modified
programs (in particular the minimizers), can now simply operate on a Code
instance and recreate a new program from that once they are finished.