Added inspection capabilities
Currently, there are two inspection modes:

History:
When enabled, Fuzzilli will store the history of all programs that are added to
the corpus in .fuzzil.history files. The history includes detailed information
about every mutations performed on the sample and includes the program's code
after every intermediate step (e.g. every execution of a mutator).

Types:
When enabled, Fuzzilli will output type all available type information for
the programs that it writes to disk in the form of comments.