Make ExpressionType a struct
In general, we should avoid having global class instances (as the ExpressionTypes
in JSExpressions.swift used to be) since that might cause performance issues
when running in multithreaded mode (e.g. due to refcounting). On the other hand,
global, constant structs should be fine (there is no refcount and they are read-only).