Simplify splicing
This change removes the ability to rewire inputs directly during splicing,
which is fairly complicated, and instead adds a TypeAwareInputMutator which
behaves like the InputMutator except that it takes type information into
account and attempts to find "compatible" inputs. The idea here is that
splicing with rewiring is roughly equivalent to a simple splicing followed by
an input mutation.