libsocket: introduce `socket_t`, adjust types (NFC) (#270)
This introduces a new `socket_t` typedef that is meant to allow handling
differences between platforms.  Windows uses a `SOCKET` type instead of
`int` for the socket handle.  Unlike Unix, not everything is a file
descriptor on Windows, but rather everything is an opaque handle, which
is word sized, and thus the pervasive use of `int` results in possible
truncation of the value.  The introduction of the typedef allows us to
lift that condition.  The remainder of the change makes the types more
pedantically correct, matching the underlying declarations on POSIX.
Surprisingly, the more pedantically correct definition makes the code
more portable.