Perform initial corpus generation when starting with an empty corpus
Instead of using a single arbitrary program as initial seed program, we
now perform initial corpus generation when starting with an empty
corpus. For that, we use a new GenerativeEngine to generate short,
random programs using the configured CodeGenerators. We do this until we
haven't found a new, interesting program for a number of iterations, at
which point we switch to the main fuzzing engine.

This change also removes the program prefix mechanism, which was used by
the MutationEngine and ProgramTemplates to insert a (again somewhat
arbitrary) chunk of code at the beginning of the generated programs.

Overall, the goal here is to make the generated programs less influenced
by arbitrary, hardcoded code snippets in the fuzzer and instead rely on
the configured CodeGenerators to generate code when needed.