Implement ExplorationMutator
Quoting from the documentation:

This mutator does the following:
  1. it inserts Explore operations for random existing variables in the program to be mutated
  2. It executes the resulting (temporary) program. The Explore operations will be lifted
    to a sequence of code that inspects the variable at runtime (using features like 'typeof' and
    'Object.getOwnPropertyNames' in JavaScript) and selects a "useful" operation to perform
    on it (e.g. load a property, call a method, ...), then reports back what it did
  3. the mutator processes the output of step 2 and replaces some of the Explore mutations
    with the concrete action that was selected at runtime. All other Explore operations are discarded.

The result is a program that performs useful actions on some of the existing variables even without
statically knowing their type. The resulting program is also deterministic and "JIT friendly" as it
no longer relies on any kind of runtime object inspection.