Remove runtime type collection support
This feature is essentially replaced by the ExplorationMutator which
also uses runtime type information but instead of forwarding it to
Fuzzilli (which then later on selects actions to perform based on the
type), it directly selects a "useful" action to perform. This is a lot
simpler and easier to maintain as the complexity is contained in a
single mutator instead of being spread throughout the codebase.