Remove FunctionSignature from FunctionDefinitions
Instead, use a simpler Parameters struct which currently just stores the
number of parameters and whether there is a rest parameter.

It likely makes little sense to encode type information about parameters
(such as "parameter 1 is a Uint8Array") in a function definition as
there is no way to ensure that the information stays accurate (e.g. due
to mutations). It's also generally not possible for a JavaScript ->
FuzzIL compiler to determine type information for parameters.

With this change, the implementation of types are again an
implementation detail and no longer exposed in e.g. protobufs.