More InliningReducer fixes and improvements
When determining whether a function can be inlined, if a LoadArguments
is seen, the function cannot be inlined. However, the LoadArguments may
also belong to e.g. a BeginMethod, which needs to be handled. In the
future, the handling of anything function-like should be unified by
using the new BeginAnySubroutine Operation class. For now though, we
need special handling until classes are refactored.