Fix handling of switch-case in JSTyper
Previously, the handling of an empty switch was wrong, as no child state
would ever be pushed, but we'd still perform a merge state operation.
With the new logic, we can now also correctly compute types for a switch
statement without a default case, in which case the types prior to the
BeginSwitch also need to be taken into account.