Implement a new Splicing Algorithm
Splicing refers to the process of taking a subset of the instructions of
one program and inserting them into another program, possibly connecting
their data-flows together. Splicing must always produce a valid program.
In particular, all input variables used by the spliced instructions must
be available, and the instructions must exist in the correct context
(e.g. a BeginSwitchCase must only occur inside a BeginSwitch).

The new splicing algorithm achieves this and runs in linear time.
Its main features include:
- Computation of data-flow dependencies between instructions to
  determine which instructions to splice
- Probabilistic remapping of some variables to existing and "compatible"
  variables in the host program to connect the dataflows of the programs
- The ability to splice instructions with special context requirements.
  For example, it can (only) splice an Await instruction if there is an
  open async function in the host program
- The ability to to splice into restricted contexts. For example, it can
  splice a SwitchCase block from one program into a Switch block in
  another program

See the many tests in ProgramBuilderTest.swift for examples and the
comments in ProgramBuilder.swift for an explanation of the algorithm.