Implement Minimization Post-Processing
The new MinimizationPostProcessor runs at the end of minimization and
attempts to add certain features - which are deemed useful for future
code mutations - back to the program as long as they don't alter the
interesting behaviour. For example, it attempts to add return statements
at the end of functions and add arguments to function calls.