Improve the ExplorationMutator's handling of crashes
Previously, we would report the instrumented program as crashing, which
will result in huge (including all of the Exploration logic) and
unreliable (as the Explore operations perform a random action) repro
cases. Now we attempt to recover the actions performed at runtime by
logging them from the JavaScript code. If successful, this will result
in short and deterministic testcases without any Explore operations. See
the comments in ExplorationMutator.swift for more details.