Implement expression inlining in the JavaScriptLifter
The lifter is now able to inline expressions, resulting in more readable
code. For example, where previously the lifter would've produced code
such as:

    const v5 = v3.a;
    const v6 = v5.b;
    const v7 = v6.c;
    const v8 = v7.d();

The lifter will now produce a single line of code:

    v3.a.b.c.d();

The inlining logic takes care of adding brackets where required and will
only inline effectful expressions if there is exactly one use and if the
execution order doesn't change. For example, the following FuzzIL code:

    v2 <- LoadProperty v1, 'x'
    v3 <- LoadProperty v1, 'y'
    v4 <- LoadProperty v1, 'z'
    v4 <- Ternary y, x, z
    Return v4

Will be lifted to:

    const v2 = v1.x;
    return v1.y ? v2 : v1.z;

And not:

    return v1.y ? v1.x : v1.z;

As the latter would access the properties in the wrong order, thereby
changing the semantics of the program.

Drive-by: change the variable prefix for certain types of variables.
Classes start with a 'C', functions now start with an 'f', and function
arguments start with an 'a':

    function f3(a4,a5) {
        return a4 + 2 * a5;
    }