Remove 'main' function from profiles
It's unclear whether there is any benefit in wrapping generated code in
a main function, and it is currently causing some issues with the
ExplorationMutator:

Consider the following program, instrumented for exploration:

  function main() {
    const explore = <set up exploration>
    function v1() {
      explore(v1, ...);
      ...;
    }
    v1();
  }

In this case, the explore operation may end up calling v1.caller, which
is the main function. This will cause 'main' to again set up and perform
the exploration, leading to duplicate exploration results and later an
error during processing in the mutator ("Invalid or duplicate ID ...").
Simply removing the 'main' functions should solve this issue.