Refactor random value generation in ProgramBuilder
This commit removes the seenXYZ sets for int/float/string values
previously seen by this builder. It is not clear that these provide any
net benefits: on the one hand, they may cause e.g. a property that was
previous set to later be read again but on the other hand they may cause
e.g. the same property or element to be defined multiple times, or the
same integer/float value to be used for multiple computations.

As an alternative, this commit now refines methods such as randIndex()
to return values from a smaller range more frequently so as for example
to increase the chance of the same element being accessed. Futhermore,
mutators such as the Exploration- or ProbingMutator also help to ensure
that existing properties are accessed or non-existant (but accessed)
properties installed.