Allow integer overflows when selecting a random element index
When adding an element to an object literal, we currently check whether
that element already exists, and if so, simply increment the index and
try again. However, it can happen that the index is Int64.max, in which
case the code would trigger an integer overflow and crash. Since it's
not clear if there's something better we could do in that case, simply
allow integer overflows in that code.