Improve placement of Explore operations
When exloring the output of a block, for example a BeginPlainFunction or
BeginClass, we should put the Explore operation directly after the end
of that block instead of as first instruction in the block's body. Doing
the latter will likely cause unchecked recursion for functions and is
syntactically invalid for classes (so there was some special casing to
prevent this, which is now removed).