Allow some actions to fail in ExplorationMutator
Consider the following example: we want to discover the following
function which is not part of the static environment model:

    function foo(o) {
        if (typeof o.someProp === 'undefined')
            throw "Invalid use";
        ...
    }

While the ExplorationMutator will find this function and attempt to call
it, the call will (almost certainly) always throw an exception since the
argument is wrong. In that case, the action was then previously
discarded and the call was never discovered.

With this change, we now allow certain actions, namely function, method,
and constructor calls to raise an exception during exploration, in which
case we'll keep the actions but surround them in try-catch. In the above
example, the ProbingMutator will then later be able to construct a valid
call to foo() as it will see the access to "someProp" and install it.