Refactor the instance synchronization protocol
The high-level protocol is now implemented in a single place:
Sync.swift, while the concrete implementations for synchronizing
instances over a network or between threads now only need to take care
of delivering the message to the other end. This makes the communication
more robust as the same messages are sent between instances no matter
how they are connected.

Moreover, this commit also refactors the corpus import mechanism, which
is now performed in small steps instead of as one large operation,
making the fuzzer responsive (UI, network, ...) during a corpus import.