Ensure that the `fuzzilli` function exists in instrumented code
If a program with one of these instrumentations causes a crash, the
crash may not be reproducible unless the `fuzzilli` function is defined.
Outside of fuzzilli-instrumented builds this function will not exist, so
this commit adds a short piece of JS code at the top of the
instrumentation code to define the function if it doesn't already exist.