Refactor while- and do-while loops in FuzzIL
These loops now have the following general shape:

    BeginLoopHeader
        // header code
    BeginLoopBody
        // body code
    EndLoop

This now allows arbitrary computations to be performed in loop headers,
as is also possible in JavaScript. It also allows us to compile all
possible while- and do-while loops from JavaScript to FuzzIL, not just
those that have a particular shape (e.g. have a single comparison, which
was previously required).

In JavaScript, a loop header must consist of exactly one expression.
However, this is purely a syntactical restriction. If the FuzzIL loop
header consists of multiple expressions, these are simply chained
together using the comma operator:

    while (foo(), bar(), baz()) {
        ...
    }

Further, if the FuzzIL loop header contains more complicated code such
as control-flow, the header is lifted to an arrow function:

    while ((() => {
        if (x) {
            return y;
        }
        return z;
    })()) {
        ...
    }