Write a settings.json file to disk.
Fuzzilli now also writes a settings.json file into the storage
directory. It is now also possible to give Fuzzilli a tag, i.e. a
revision which is also stored in the settings.json file. This makes it
easy to associate a given storage directory with the revision that was
fuzzed and the used command line arguments.