Fix bug in NetworkSync
In case of an error on the connection from a child to a parent node, it
can happen that Connection.error is called multiple times: the first
time for the actual error, and then again in case there is any pending
data to send out (which will cause the "Attempted to send data on a
closed connection" error). However, the handler's handleError method
should still only be called once, for the original error, as it may
otherwise cause the handler (NetworkChild.Transport) to attempt to
reconnect to the parent node a second time, which will cause other
problems.