Fix OperationMutator's handling of CreateTemplateString
The OperationMutator can replace a random part of a template literal
with a different string. Previously, this was done using the
replaceRandomElement helper function. That, however, converted the array
of the string parts into a set, which removed any duplicate entries and
therefore changed the number of parts. This in turn would lead to an
inconsistent instruction as the number of inputs expected by the
Operation no longer matched the number of inputs in the Instruction.