Remove reuseOrLoad APIs
These were originally introduced to reduce "variable pollution", i.e.
having many uninteresting variables in the active scope.
For integers they are no longer used now after the loop refactoring
because the loop counts are created in the loop header scopes and so are
not visible outside (and so it doesn't matter if they are new variables
or reuse existing ones). For builtins, it's probably not worth the
effort to try and deduplicate them. Instead, we should probably change
the variable selection heuristics to prefer newer variables so that
the result of whatever is done with the builtin is preferred for
following operations instead of the builtin itself.