Implement DeduplicatingReducer
This is a simple minimization pass that attempts to deduplicate
variables holding identical values. Currently it only deduplications
variables from LoadBuiltin operations. For example:

    v0 <- LoadBuiltin 'foo'
    v1 <- CallFunction v0
    ...
    v23 <- LoadBuiltin 'foo'
    v24 <- CallFunction v23

Will be turned into

    v0 <- LoadBuiltin 'foo'
    v1 <- CallFunction v0
    ...
    v23 <- CallFunction v0

It could be extended to cover other types of values as well, however,
it's not obvious that that is beneficial. For example, if identical
string or number values are kept in separate variables, they can be
mutated independently, which may be interesting. In contrast, it's
probably less useful to mutate the builtin that is accessed by a
LoadBuiltin operation as that will often result in invalid or
meaningless code (e.g. a nonexistant property or method being used).