Implement support for optional chaining
This is based on pull request #322

We now have a `isGuarded` flag in various property-related operations
(GetXYZ, DeleteXYZ, and CallXYZMethod) which, if enabled, causes the
lifter to use the `?.` (and equivalent) operators instead of `.`. This
makes the property access guarded in the sense that it will not cause a
runtime exception if the access object is "nullish" (undefined or null).