Introduce ValueGenerators
ValueGenerators are regular CodeGenerators that can additionally be used
to bootstrap code generation by producing a number of new variables
without requiring any existing variables as inputs.
This new mechanism replaces the previous "trivialCodeGenerators"
mechanism. The main benefit of the new approach is that the value
generators (and therefore the available variables at the start of code
generation) are controllable through the code generators weights.