Implement ImitationGenerator
This code generator will create a new object that "imitates" an existing
value. For example, for primitive values this will generate a new object
with a valueOf/@@toPrimitive callback, and for objects it may create a
new object with the original object as prototype.
The imitation object can then be used wherever the original value may
have been used, but will (probably) behave in subtly (or less subtly)
different ways, for example by executing arbitrary code when used.