Fix InliningReducer after introducing BeginObjectLiteralComputedMethod
It can now happen that we have a function that is referenced as input to
a CallFunction operation, but also as input to a
BeginObjectLiteralComputedMethod. In that case, we previously did not
"see" the 2nd use of the function variable and would (incorrectly)
attempt to inline the function. As that cannot work, we would then fail
an assertion later on. With this change, we now realize that the
function has more than one use and will not attempt to inline it.