Remove the .unknown type
Instead, we now use .anything to express that the type is unknown, which
is more natural. It also allows us to remove some hacky code that was
previously necessary to ensure that variable types were .unknown while
input types (e.g. in function signatures) were .anything if the type
isn't known.

Previously, when querying variables we explicitly also looked for
variables of .unknown type to ensure that such variables would also be
used. Now, instead of that we simply perform the type query with .MayBe
instead of .Is in aggressive mode, which achieves the same thing
(.anything MayBe any other type) but is more general: we are now also
able to for example return variables of type .number (an integer or a
float) when requesting a .integer, etc.