Remove type generalization
Previously, we had code such as

    let queriedType = b.type(of: v).generalize()
    let replacement = b.randomVariable(ofType: queriedType)

The goal here was to find a type that may be compatible. For example, we
may generalize a function with a specific signature to a generic
function, then use that to effectively search for any function. In pseudocode:

    let generalizedFunction = functionType.generalize()
    let replacement = b.findVariable(where: { $0.Is(generalizedFunction) })

However, generalization is not well defined, and we should instead use
the (well-defined) MayBe query to find a replacement that may be
compatible (and the Is query to find a replacement that definitiely is
compatible). The pseudo-code above then simply becomes:

    let replacement = b.findVariable(where: { $0.MayBe(funcType) })

This covers the scenario above and other scenarios that weren't handled
previously.

Since randomVariable(ofType: T) uses a MayBe query in aggressive mode
and an Is query in conservative mode, we get fairly nice results now.
For example, when splicing in aggressive mode, we may now replace one
function with a different one as we used to do before. However, in
conservative mode we will only replace the functions if they are known
to be compatible, such as e.g. have the same number of parameters of
compatible types, etc.

This change required implementing function signature subsumption, which
is non-trivial. See the many added test cases for examples of the
subsumption rules for function signatures.