Refactor the randomVariable() variants in ProgramBuilder
We used to have:
    - randomVariable(ofType:)
    - randomVariable(ofConservativeType:)

Now we have:
    - randomVariable(forUseAs:)
    - randomVariable(ofType:)

which behave the same as their former counterpart:
randomVariable(forUseAs:) will find a variable that may be suitable for
use as the given type. This means that it may return variables of a
broader type (for example, a variable of type .number or even .anything
if a .integer is requested). In this way, this function ensures that
also variables for which no exact type could be statically inferred will
still be used as inputs for following code. On the other hand,
randomVariable(ofType:) will always return a variable that is known to
have the requested type (i.e. for which `b.type(of: v).Is(type)` is
true. The new names should hopefully be more intuitive.