Remove ProgramBuilder.generateVariable()
Previously, generateVariable() could be used to try and generate a
variable of a given type, which was also used by things like
generateCallArguments().

However, this is most likely the wrong approach. Instead, code
generation should rely exclusively on CodeGenerators to generate
instructions so that the weights of the generators fully control the
"shape" of the generated code. ProgramBuilder.generateVariable() could
not cover all possible ways of generating, e.g., an object, and so it
would bias the generated code. Further, using helper functions such as
generateCallArguments() would previously often lead to a situation where
the first call to a function generated argument values, and all
subsequent calls to that function simply reused those argument values,
which is also not ideal.

Now, instead of trying to generate variables of specific types, we
should always use existing variables, and if there are no existing
variables of the desired type, we should use another random variable but
guard against runtime exceptions. This way, we will still be able to
perform the operations that we intended to perform and also allow for
future mutations (such as the ProbingMutator) to fix the inputs to make
the operation fully valid. To increase the diversity of e.g. arguments
used for funtion calls, we can add simple rules to for example return a
random variable if less than N variables of the specified type are
currently available.