Add ProgramBuilder.buildPrefix()
This method can now be used to generate some prefix code when building
new programs. It will use ValueGenerators to generate a number of
variables with known types which can then be used by the following code.

We now also forbid calling .build() when there are no visible variables
currently available as that was previously handled only with a hacky
workaround. Instead, we now force client code to first explicitly run
buildPrefix() before calling build(). This way, when calling for example
build(n: 10) it is clear that this budget will be completely used for
"real" code generation or splicing and not end up being spent for value
generation, which is less interesting.