Replace generateSignature() with randomParameters()
Previously, when wanting to generate a subroutine with parameters of
type other than .anything, one had to use generateSignature(), which
would generate some completely random types and construct a signature
from that. The main issue with that approach is that it makes it
unlikely that we'll have matching argument values for calling the
function later on.

Instead, we now use ProgramBuilder.randomParameters() everywhere, which
this commit also improves to allow it to return parameters of types
other than .anything. The crucial difference is that randomParameters()
will pick parameter types based on types that are _already present_ in
the current program, so that it is guaranteed that matching argument
values will later be available. For example, if there are a handful of
.integer values present in the current program, .randomParameters() may
generate parameters of type .integer so that the existing variables can
later be used to call the function. See the added tests for more
examples of the new .randomParameters() functionality.