Implement variable hiding in the ProgramBuilder
This feature now allows users of a ProgramBuilder to explicitly hide a
variable, in which case it will not be returned from `randomVariable()`
and related functions and will not show up in the `visibleVariables`
list until it is unhidden again. This can be used by CodeGenerators that
need to create temporary variables which should not be used by following
CodeGenerators. The feature is now also used to implement a "recursion
guard" feature: when a new function is created, the variable containing
the function itself is hidden inside the function's body so that trivial
recursion is not possible. This now replaces the (fairly crude) former
mechanism that attempted to prevent recursion.

With this change, the scope management is moved out of the
VariableAnalyzer and is now performed directly by the ProgramBuilder.
This may be useful in the future to implement faster lookup of variables
by types.