Support RecursiveValueGenerators
These mostly behave as expected: they are ValueGenerators so can run
without visible variables and will always create some (with statically
inferrable types), but they are also recursive and may call back into
the code building machinery, for example to fill the bodies of newly
created blocks.

In practice, we now use this for object- and class literal building
by ValueGenerators.

One subtle consequence of this is that it is no longer guaranteed that,
after running _multiple_ value generators, that all variables have a
known type. That is now only guaranteed for a single ValueGenerator. The
reason is that recursive code building may generate reassignments of
existing variables, which can change the types of variables created by
previous ValueGenerators to .anything.