Add Operation.Attributes.isSingular
This attribute expresses that this type of operation should only occur
once in its surrounding block. Examples include default switch cases or
class constructors.
If there are multiple singular operations inside the same block, then
all but the first one are ignored. This way of handling singular
operations is easy to implement: in the JavaScriptLifter, we simply
comment out all singular operations/blocks after the first one. It would
also be possible to forbid multiple such operations completely, but that
is more complicated to enforce and/or would mean that singular
operations can never be created through code generators or splicing, but
only when the surrounding block is initially created.