Introduce GuardableOperation
A GuardableOperation is one that can be guarded against runtime
exceptions caused by it. If guarding is active, any exception raised
will be suppressed and following code will execute normally.

When lifting to JavaScript, guarding is usually implemented by wrapping
the code in a try-catch, but special handling is also possible. For
example, guarded property loads use optional chaining: `o?.a` instead.

This can now be used by CodeGenerators if they aren't sure that the
operation will not cause a runtime exception. For example, the function
call generator now looks like this:

    let arguments = b.randomArguments(forCalling: f)
    let needGuard = b.type(of: f).MayNotBe(.function())
    b.callFunction(f, withArgs: arguments, guard: needGuard)

Similarly, the method call generator will use guarding if it isn't sure
that the method exists. In the future, various other operations could be
guarded, such as binary operations (if mixing bigints and numbers),
property stores (if the input may be nullish), or destruct operations
(if the input isn't iterable).

There are a few benefits to adding a language feature for guarding
instead of, for example, creating explicit try-catch blocks:
1. both the generated code and the CodeGenerators become shorter.
2. it allows outputs of guarded instructions to be used as inputs for
   subsequent instructions as they remain in scope.
3. it makes it trivial to turn guarded operations into unguarded
   operations during minimization.
4. it makes it easy to implement runtime "fixing" of guarded operations
   to determine if the guard is needed, and remove it if it isn't.