Wasm Fuzzing Foundation
This CL is the foundation of a new Part of the FuzzIL, a part that shows
how we can leverage the powerful FuzzIL to lift into a different target
language (i.e. Wasm) that is unified through Fuzzilli's powerful type
system. This allows us to fuzz the full V8 engine from JS to Wasm and
back again. We maintain the good properties of FuzzIL that lends itself
to mutation based fuzzing and compile it down to Wasm. Fuzzilli can
maintain strict typing that is necessary in Wasm while being flexible
enough to do powerful mutations in the JS parts.