Improve the MapTransitionTemplate
Previously we were manually tracking the live variables that contain
objects created by the template and then used that list to call
%HeapObjectVerify on all visible objects at the end of the program. This
is somewhat fragile however, and essentially redundant work since the
ProgramBuilder also has full knowledge of all visible variables. So now
we only rely on the .visibleVariables exposed by the builder.

Further, the custom code generators used by the template previously
failed to ensure that their input types were of the expected type, which
would quickly lead to runtime exceptions.

Finally, this change also fixes a small issue in the JSTyper which
failed to correctly type subclass instance types if the return value of
the super constructor was not an object. See the added tests.