Introduce CodeGenerator.Inputs
Currently, the inputTypes of code generators are only "hints": the
ProgramBuilder uses randomVariable(forUseAs:) to find matching variables
which can return variables of wider or even different types (in
particular, it can return variables of unknown types so that these are
used as inputs as well). This behavior is usually what we want, but some
code generators can only do meaningful work if they get the "right"
input, for example a function that can be JIT compiled by calling it
repeatedly.
This change now introduces a new struct, CodeGenerator.Inputs, which
contains the input types together with an InputMode which now makes this
behavior explicit. The default mode is .loose, in which case the
behavior is the same as before, allowing also variables of unknown types
to be used as inputs. The .strict mode on the other hand indicates that
the code generator must be called with inputs of the specified types,
and so the ProgramBuilder uses randomVariable(ofType:) to find matching
inputs.
When CodeGenerators are defined, they can express which mode they want
by using convenience constructors of the Inputs struct. For example,

    input: .one

would result in one input of type .anything

    input: .preferred(.object())

would result in one input of type .object() and mode .loose, and

    input .required(.function(), .integer)

would result in two inputs and mode .strict.